# server/openapi_spec.yaml
openapi: 3.0.0
info:
  title: Astronova Backend API
  description: Comprehensive astrology API with AI-powered insights, birth charts, compatibility matching, and personalized reports
  version: 1.0.0
  contact:
    name: Astronova Support
    email: admin@100xai.engineering
servers:
  - url: https://astronova-ghcr.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: "Admin API token. Production admin endpoints also accept Authorization: Bearer <ADMIN_API_TOKEN>."
...
paths:
...
  /api/v1/admin/grant-pro:
    post:
      summary: Grant pro subscription to a user
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pro subscription granted
...
  /api/v1/admin/list-users:
    get:
      summary: List all users
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
...